[WIP] switch to user mods

+++ scripts/setup_ctsm.sh
@@ -6,4 +6,7 @@ export PYTHONPATH=$(dirname $(dirname $(realpath $0)))
 
 python app/utils/dependencies.py
 
-rsync -rv resources/overwrites/ resources/ctsm
+# If resources/overwrites/ exists, run rsync
+if [[ -d resources/overwrites ]]; then
+  rsync -rv resources/overwrites/ resources/ctsm
+fi
